:+1: Move implementation into `impl.ts` to reduce the size
All denops plugin refer `mod.ts` directly/indirectly means that all denops plugins
refer `denops.ts` indirectly thus the size of `denops.ts` should be as small as possible.
