+++ denops/@denops-private/worker/script.ts
@@ -13,7 +13,7 @@ import {
 } from "https://deno.land/x/workerio@v1.4.3/mod.ts#^";
 import { responseTimeout } from "../defs.ts";
 import type { Denops, Meta } from "../../@denops/mod.ts";
-import { DenopsImpl } from "../../@denops/denops.ts";
+import { DenopsImpl } from "../../@denops/impl.ts";
 
 const worker = self as unknown as Worker;
