+++ denops/@denops/denops.ts
@@ -1,5 +1,3 @@
-import { Session } from "https://deno.land/x/msgpack_rpc@v3.1.4/mod.ts#^";
-
 /**
  * Method dispatcher
  */
@@ -109,76 +107,4 @@ export interface Denops {
    * @param args: Arguments of the function.
    */
   dispatch(name: string, fn: string, ...args: unknown[]): Promise<unknown>;
-}
-
-function normArgs(args: unknown[]): unknown[] {
-  const normArgs = [];
-  for (const arg of args) {
-    if (arg === undefined) {
-      break;
-    }
-    normArgs.push(arg);
-  }
-  return normArgs;
-}
-
-export class DenopsImpl implements Denops {
-  readonly name: string;
-  readonly meta: Meta;
-  #session: Session;
-
-  constructor(
-    name: string,
-    meta: Meta,
-    session: Session,
-  ) {
-    this.name = name;
-    this.meta = meta;
-    this.#session = session;
-  }
-
-  get dispatcher(): Dispatcher {
-    return this.#session.dispatcher;
-  }
-
-  set dispatcher(dispatcher: Dispatcher) {
-    this.#session.dispatcher = dispatcher;
-  }
-
-  async call(fn: string, ...args: unknown[]): Promise<unknown> {
-    return await this.#session.call("call", fn, ...normArgs(args));
-  }
-
-  async batch(
-    ...calls: [string, ...unknown[]][]
-  ): Promise<unknown[]> {
-    const normCalls = calls.map(([fn, ...args]) => [fn, ...normArgs(args)]);
-    const [results, errmsg] = await this.#session.call(
-      "batch",
-      ...normCalls,
-    ) as [
-      unknown[],
-      string,
-    ];
-    if (errmsg !== "") {
-      throw new BatchError(errmsg, results);
-    }
-    return results;
-  }
-
-  async cmd(cmd: string, ctx: Context = {}): Promise<void> {
-    await this.#session.call("call", "denops#api#cmd", cmd, ctx);
-  }
-
-  async eval(expr: string, ctx: Context = {}): Promise<unknown> {
-    return await this.#session.call("call", "denops#api#eval", expr, ctx);
-  }
-
-  async dispatch(
-    name: string,
-    fn: string,
-    ...args: unknown[]
-  ): Promise<unknown> {
-    return await this.#session.call("dispatch", name, fn, ...args);
-  }
 }
